revert bluetooth-serial-port stuff
ok seriously, i hate nodejs so much
